finalize
formatting

fix

f

f

+++ db/migrate/20210712122757_create_discount_rules.rb
@@ -2,7 +2,7 @@ class CreateDiscountRules < ActiveRecord::Migration[6.1]
   def change
     create_table :discount_rules do |t|
       t.string :discount_type, null: false
-      t.integer :amount, null: false
+      t.float :amount, null: false
       t.integer :value, null: false
 
       t.timestamps
